Add responsiveness to new navbar and clean up code
