+++ package.json
@@ -13,14 +13,14 @@
     "@aws-amplify/auth": "^6.13.0",
     "@aws-amplify/ui-react": "^6.11.1",
     "@headlessui/react": "^2.2.1",
+    "@hookform/resolvers": "^5.1.1",
+    "@preline/toggle-password": "^3.1.0",
     "@radix-ui/react-avatar": "^1.1.10",
     "@radix-ui/react-tabs": "^1.1.12",
+    "aws-amplify": "^6.15.0",
     "bad-words": "^4.0.0",
     "class-variance-authority": "^0.7.1",
     "date-fns": "^4.1.0",
-    "@hookform/resolvers": "^5.1.1",
-    "@preline/toggle-password": "^3.1.0",
-    "aws-amplify": "^6.15.0",
     "lucide-react": "^0.508.0",
     "next": "15.2.4",
     "openid-client": "^6.5.0",
@@ -39,8 +39,8 @@
     "@aws-sdk/client-cognito-identity-provider": "^3.828.0",
     "@eslint/eslintrc": "^3",
     "@tailwindcss/postcss": "^4",
-    "@types/date-fns": "^2.5.3",
     "@types/aws-lambda": "^8.10.149",
+    "@types/date-fns": "^2.5.3",
     "@types/node": "^20",
     "@types/react": "^19",
     "@types/react-dom": "^19",
