+++ src/app/page.tsx
@@ -5,15 +5,13 @@ import AboutSection from "@/components/home/about-section"; // Adjust the path a
 import FeaturesSection from "@/components/home/feature-section";
 import { MeetOurProviders } from "@/components/home/meet-provider-section";
 import { HowItWorks } from "@/components/home/how-it-works";
-import NavBar from "@/components/navbars/navbar";
+import NavBar from "@/components/nav-bars/navbar";
 
 export default async function Home() {
   return (
     <div>
       <main>
-        {/* Temp location of top navbar */}
-        {/* <TopNavBar /> */}
-        {/* <NavBar isSignedIn={await isAuthenticated()} /> */}
+        {/* Temp location of navbar */}
         <NavBar />
         <HeroSection />
         <AboutSection />
