+++ src/components/nav-bars/registration-nav-sidebar.tsx
@@ -2,6 +2,7 @@
 
 interface RegistrationNavSideBarProps {
   activeSection: string;
+  // eslint-disable-next-line no-unused-vars
   onSectionClick: (sectionId: string) => void;
   sectionCompletion: Record<string, { completed: boolean; progress: number }>;
 }
