--- src/components/signup-forms/signup-form.tsx
@@ -1,21 +1,0 @@
-"use client";
-
-//import { signUp } from "aws-amplify/auth";
-import { AuthUser } from "aws-amplify/auth";
-import { useEffect } from "react";
-import { redirect } from "next/navigation";
-import { withAuthenticator } from "@aws-amplify/ui-react";
-import "@aws-amplify/ui-react/styles.css";
-
-const SignUpForm = ({ user }: { user?: AuthUser }) => {
-  //redirect user to home if logged in
-  useEffect(() => {
-    if (user) {
-      redirect("/");
-    }
-  });
-
-  return null;
-};
-
-export default withAuthenticator(SignUpForm);
